pyproject.toml
[project]
name = "mimir-poc"
version = "0.1.0"
description = "Proof of concept for Mimir travel companion app"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-vision>=3.9.0",
    "pydantic-ai>=0.0.21",
    "pydantic>=2.10.6",
    "streamlit>=1.41.1",
    "pillow>=11.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.4",
]
